Improve and refactor code
