+++ apps/rebalancer/src/rebalance/brla-to-axlusdc/steps.ts
@@ -5,8 +5,6 @@ import {
   ApiManager,
   AveniaFeeType,
   AveniaPaymentMethod,
-  AveniaSwapTicket,
-  AveniaTicketStatus,
   BrlaApiService,
   BrlaCurrency,
   checkEvmBalancePeriodically,
@@ -34,41 +32,10 @@ import {
 import Big from "big.js";
 import { encodeFunctionData } from "viem";
 import { polygon } from "viem/chains";
-import { brlaFiatTokenDetails, brlaMoonbeamTokenDetails, usdcTokenDetails } from "../../constants.ts";
+import { brlaMoonbeamTokenDetails, usdcTokenDetails } from "../../constants.ts";
+import { checkTicketStatusPaid } from "../../utils/brla.ts";
 import { getConfig, getMoonbeamEvmClients, getPendulumAccount, getPolygonEvmClients } from "../../utils/config.ts";
 import { waitForTransactionConfirmation } from "../../utils/transactions.ts";
-
-async function checkTicketStatusPaid(brlaApiService: BrlaApiService, ticketId: string): Promise<AveniaSwapTicket> {
-  const pollInterval = 5000; // 5 seconds
-  const timeout = 5 * 60 * 1000; // 5 minutes
-  const startTime = Date.now();
-  let lastError: any;
-
-  while (Date.now() - startTime < timeout) {
-    try {
-      const ticket = await brlaApiService.getAveniaSwapTicket(ticketId);
-      if (ticket && ticket.status) {
-        if (ticket.status === AveniaTicketStatus.PAID) {
-          return ticket;
-        }
-        if (ticket.status === AveniaTicketStatus.FAILED) {
-          throw new Error("Ticket status is FAILED");
-        }
-      }
-    } catch (error) {
-      lastError = error;
-      console.warn(`Polling for ticket ${ticketId} status failed with error. Retrying...`, lastError);
-    }
-    await new Promise(resolve => setTimeout(resolve, pollInterval));
-  }
-
-  if (lastError) {
-    console.error("Polling for ticket status timed out with an error: ", lastError);
-    throw new Error(`Polling for ticket status timed out with an error: ${lastError.message}`);
-  }
-
-  throw new Error("Polling for ticket status timed out.");
-}
 
 export async function checkInitialPendulumBalance(pendulumAddress: string, requiredAmount: string): Promise<Big> {
   const apiManager = ApiManager.getInstance();
